+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
   * `ecs_domain_name` to `ecs_domain_extension`
   * `hms_readonly_instance_count` to `hms_ro_ecs_task_count`
   * `hms_readwrite_instance_count` to `hms_rw_ecs_task_count`
+- Optimize ECS task S3 policy.
 
 ### Added
 - Option to send Hive Metastore metrics to CloudWatch - see [#4](https://github.com/ExpediaInc/apiary-metastore-docker/issues/4).